materialsandproducts

+++ SistemaFerredomos/src/Models/ProductsModel.cs
@@ -10,7 +10,7 @@ namespace SistemaFerredomos.src.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Stock { get; set; }
+        public int Stock { get; set; }
         public decimal PurchasePrice { get; set; }
         public decimal SalePrice { get; set; }
         public int SupplierId { get; set; }
